Remove redundant xmr_provider.py class
