+++ anydex/test/wallet/monero/__init__.py
@@ -1,0 +1,3 @@
+"""
+Contains tests for the Monero wallet module.
+"""
